Initial pass on cleanup
