+++ europi.go
@@ -1,57 +1,55 @@
-package europi // import europi "github.com/awonak/EuroPiGo"
+package europi // import "github.com/heucuva/europi"
 
 import (
 	"machine"
-)
 
-const (
-	MaxVoltage = 10.0
-	MinVoltage = 0.0
+	"github.com/heucuva/europi/input"
+	"github.com/heucuva/europi/output"
 )
 
 // EuroPi is the collection of component wrappers used to interact with the module.
 type EuroPi struct {
 	// Display is a wrapper around ssd1306.Device
-	Display *Display
+	Display *output.Display
 
-	DI DigitalReader
-	AI AnalogReader
+	DI input.DigitalReader
+	AI input.AnalogReader
 
-	B1 DigitalReader
-	B2 DigitalReader
+	B1 input.DigitalReader
+	B2 input.DigitalReader
 
-	K1 AnalogReader
-	K2 AnalogReader
+	K1 input.AnalogReader
+	K2 input.AnalogReader
 
-	CV1 Outputer
-	CV2 Outputer
-	CV3 Outputer
-	CV4 Outputer
-	CV5 Outputer
-	CV6 Outputer
-	CV  [6]Outputer
+	CV1 output.Output
+	CV2 output.Output
+	CV3 output.Output
+	CV4 output.Output
+	CV5 output.Output
+	CV6 output.Output
+	CV  [6]output.Output
 }
 
 // New will return a new EuroPi struct.
 func New() *EuroPi {
-	cv1 := NewOutput(machine.GPIO21, machine.PWM2)
-	cv2 := NewOutput(machine.GPIO20, machine.PWM2)
-	cv3 := NewOutput(machine.GPIO16, machine.PWM0)
-	cv4 := NewOutput(machine.GPIO17, machine.PWM0)
-	cv5 := NewOutput(machine.GPIO18, machine.PWM1)
-	cv6 := NewOutput(machine.GPIO19, machine.PWM1)
+	cv1 := output.NewOutput(machine.GPIO21, machine.PWM2)
+	cv2 := output.NewOutput(machine.GPIO20, machine.PWM2)
+	cv3 := output.NewOutput(machine.GPIO16, machine.PWM0)
+	cv4 := output.NewOutput(machine.GPIO17, machine.PWM0)
+	cv5 := output.NewOutput(machine.GPIO18, machine.PWM1)
+	cv6 := output.NewOutput(machine.GPIO19, machine.PWM1)
 
-	return &EuroPi{
-		Display: NewDisplay(machine.I2C0, machine.GPIO0, machine.GPIO1),
+	e := &EuroPi{
+		Display: output.NewDisplay(machine.I2C0, machine.GPIO0, machine.GPIO1),
 
-		DI: NewDI(machine.GPIO22),
-		AI: NewAI(machine.ADC0),
+		DI: input.NewDigital(machine.GPIO22),
+		AI: input.NewAnalog(machine.ADC0),
 
-		B1: NewButton(machine.GPIO4),
-		B2: NewButton(machine.GPIO5),
+		B1: input.NewButton(machine.GPIO4),
+		B2: input.NewButton(machine.GPIO5),
 
-		K1: NewKnob(machine.ADC1),
-		K2: NewKnob(machine.ADC2),
+		K1: input.NewKnob(machine.ADC1),
+		K2: input.NewKnob(machine.ADC2),
 
 		CV1: cv1,
 		CV2: cv2,
@@ -59,6 +57,8 @@ func New() *EuroPi {
 		CV4: cv4,
 		CV5: cv5,
 		CV6: cv5,
-		CV:  [6]Outputer{cv1, cv2, cv3, cv4, cv5, cv6},
+		CV:  [6]output.Output{cv1, cv2, cv3, cv4, cv5, cv6},
 	}
+
+	return e
 }
